FLUME-3148 Make tests run in parallel
updating surefire plugin, setting fork options, adding a testPortProvider
