+++ flume-ng-channels/flume-kafka-channel/pom.xml
@@ -59,6 +59,11 @@ limitations under the License.
       <classifier>tests</classifier>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
 
   </dependencies>
