+++ flume-ng-clients/flume-ng-log4jappender/pom.xml
@@ -83,6 +83,12 @@ limitations under the License.
 
     <dependency>
       <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.flume</groupId>
       <artifactId>flume-ng-configuration</artifactId>
       <scope>test</scope>
     </dependency>
