+++ flume-ng-clients/flume-ng-log4jappender/src/test/java/org/apache/flume/clients/log4jappender/TestLog4jAppenderWithAvro.java
@@ -48,6 +48,7 @@ import org.apache.flume.channel.MemoryChannel;
 import org.apache.flume.channel.ReplicatingChannelSelector;
 import org.apache.flume.conf.Configurables;
 import org.apache.flume.source.AvroSource;
+import org.apache.flume.test.util.TestPortProvider;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.apache.log4j.PropertyConfigurator;
@@ -59,6 +60,7 @@ public class TestLog4jAppenderWithAvro {
   private AvroSource source;
   private Channel ch;
   private Properties props;
+  private int port;
 
   @Before
   public void setUp() throws Exception {
@@ -66,7 +68,7 @@ public class TestLog4jAppenderWithAvro {
     Files.copy(Resources.newInputStreamSupplier(schemaUrl),
         new File("/tmp/myrecord.avsc"));
 
-    int port = 25430;
+    port = TestPortProvider.getInstance().getFreePort();
     source = new AvroSource();
     ch = new MemoryChannel();
     Configurables.configure(ch, new Context());
@@ -99,6 +101,7 @@ public class TestLog4jAppenderWithAvro {
   @Test
   public void testAvroGeneric() throws IOException {
     loadProperties("flume-log4jtest-avro-generic.properties");
+    props.put("log4j.appender.out2.Port", String.valueOf(port));
     PropertyConfigurator.configure(props);
     Logger logger = LogManager.getLogger(TestLog4jAppenderWithAvro.class);
     String msg = "This is log message number " + String.valueOf(0);
@@ -136,6 +139,7 @@ public class TestLog4jAppenderWithAvro {
   @Test
   public void testAvroReflect() throws IOException {
     loadProperties("flume-log4jtest-avro-reflect.properties");
+    props.put("log4j.appender.out2.Port", String.valueOf(port));
     PropertyConfigurator.configure(props);
     Logger logger = LogManager.getLogger(TestLog4jAppenderWithAvro.class);
     String msg = "This is log message number " + String.valueOf(0);
@@ -173,6 +177,7 @@ public class TestLog4jAppenderWithAvro {
   @Test
   public void testDifferentEventTypesInBatchWithAvroReflect() throws IOException {
     loadProperties("flume-log4jtest-avro-reflect.properties");
+    props.put("log4j.appender.out2.Port", String.valueOf(port));
     PropertyConfigurator.configure(props);
     Logger logger = LogManager.getLogger(getClass());
     List<Object> events = Arrays.asList("string", new AppEvent("appEvent"));
@@ -203,6 +208,7 @@ public class TestLog4jAppenderWithAvro {
   @Test
   public void testDifferentEventTypesInBatchWithAvroGeneric() throws IOException {
     loadProperties("flume-log4jtest-avro-generic.properties");
+    props.put("log4j.appender.out2.Port", String.valueOf(port));
     PropertyConfigurator.configure(props);
     Logger logger = LogManager.getLogger(getClass());
     String msg = "Avro log message";
