+++ flume-ng-clients/flume-ng-log4jappender/src/test/resources/flume-loadbalancing-backoff-log4jtest.properties
@@ -15,7 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 log4j.appender.out2 = org.apache.flume.clients.log4jappender.LoadBalancingLog4jAppender
-log4j.appender.out2.Hosts = localhost:25430 localhost:25431 localhost:25432
+log4j.appender.out2.Hosts = #random from java code
 log4j.appender.out2.Selector = ROUND_ROBIN
 log4j.appender.out2.MaxBackoff = 30000
 log4j.logger.org.apache.flume.clients.log4jappender = DEBUG,out2
