+++ flume-ng-clients/flume-ng-log4jappender/src/test/resources/flume-loadbalancing-rnd-log4jtest.properties
@@ -15,6 +15,6 @@
 # specific language governing permissions and limitations
 # under the License.
 log4j.appender.out2 = org.apache.flume.clients.log4jappender.LoadBalancingLog4jAppender
-log4j.appender.out2.Hosts = localhost:25430 localhost:25431 localhost:25432 localhost:25433 localhost:25434 localhost:25435 localhost:25436 localhost:25437 localhost:25438 localhost:25439
+log4j.appender.out2.Hosts = #random from java code
 log4j.appender.out2.Selector = RANDOM
 log4j.logger.org.apache.flume.clients.log4jappender = DEBUG,out2
