+++ flume-ng-clients/flume-ng-log4jappender/src/test/resources/flume-loadbalancinglog4jtest.properties
@@ -15,5 +15,5 @@
 # specific language governing permissions and limitations
 # under the License.
 log4j.appender.out2 = org.apache.flume.clients.log4jappender.LoadBalancingLog4jAppender
-log4j.appender.out2.Hosts = localhost:25430 localhost:25431
+log4j.appender.out2.Hosts = #random from java code
 log4j.logger.org.apache.flume.clients.log4jappender = DEBUG,out2
