+++ flume-ng-clients/flume-ng-log4jappender/src/test/resources/flume-log4jtest-avro-reflect.properties
@@ -15,7 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 log4j.appender.out2 = org.apache.flume.clients.log4jappender.Log4jAppender
-log4j.appender.out2.Port = 25430
+log4j.appender.out2.Port = #random from java code
 log4j.appender.out2.Hostname = localhost
 log4j.appender.out2.AvroReflectionEnabled = true
 log4j.logger.org.apache.flume.clients.log4jappender = DEBUG,out2
