+++ flume-ng-core/pom.xml
@@ -83,6 +83,13 @@ limitations under the License.
             </lifecycleMappingMetadata>
           </configuration>
         </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-surefire-plugin</artifactId>
+          <configuration>
+            <reuseForks>false</reuseForks>
+          </configuration>
+        </plugin>
       </plugins>
     </pluginManagement>
 
@@ -380,6 +387,11 @@ limitations under the License.
     <dependency>
       <groupId>org.apache.mina</groupId>
       <artifactId>mina-core</artifactId>
+    </dependency>
+
+    <dependency>
+        <groupId>org.apache.flume</groupId>
+        <artifactId>flume-test-common</artifactId>
     </dependency>
 
   </dependencies>
