+++ flume-ng-core/src/test/java/org/apache/flume/source/TestAvroSource.java
@@ -54,6 +54,7 @@ import org.apache.flume.lifecycle.LifecycleState;
 import org.apache.flume.source.avro.AvroFlumeEvent;
 import org.apache.flume.source.avro.AvroSourceProtocol;
 import org.apache.flume.source.avro.Status;
+import org.apache.flume.test.util.TestPortProvider;
 import org.jboss.netty.channel.ChannelPipeline;
 import org.jboss.netty.channel.socket.SocketChannel;
 import org.jboss.netty.channel.socket.nio.NioClientSocketChannelFactory;
@@ -71,14 +72,15 @@ public class TestAvroSource {
   private static final Logger logger = LoggerFactory
       .getLogger(TestAvroSource.class);
 
-  private int selectedPort;
   private AvroSource source;
   private Channel channel;
   private InetAddress localhost;
+  private TestPortProvider testPortProvider;
 
   @Before
   public void setUp() throws UnknownHostException {
     localhost = InetAddress.getByName("127.0.0.1");
+    testPortProvider = TestPortProvider.getInstance();
     source = new AvroSource();
     channel = new MemoryChannel();
 
@@ -95,25 +97,13 @@ public class TestAvroSource {
 
   @Test
   public void testLifecycle() throws InterruptedException {
-    boolean bound = false;
-
-    for (int i = 0; i < 100 && !bound; i++) {
-
-      Context context = new Context();
-
-      context.put("port", String.valueOf(selectedPort = 41414 + i));
-      context.put("bind", "0.0.0.0");
-      // Invalid configuration may throw a FlumeException which has to be expected in the callers
-      Configurables.configure(source, context);
-      try {
-        source.start();
-        bound = true;
-      } catch (FlumeException e) {
-        /*
-         * NB: This assume the failure is to bind.
-         */
-      }
-    }
+    Context context = new Context();
+    int selectedPort = testPortProvider.getFreePort();
+    context.put("port", String.valueOf(selectedPort));
+    context.put("bind", "0.0.0.0");
+    // Invalid configuration may throw a FlumeException which has to be expected in the callers
+    Configurables.configure(source, context);
+    source.start();
 
     Assert
         .assertTrue("Reached start or error", LifecycleController.waitForOneOf(
@@ -132,7 +122,7 @@ public class TestAvroSource {
   public void testSourceStoppedOnFlumeExceptionIfPortUsed()
       throws InterruptedException, IOException {
     final String loopbackIPv4 = "127.0.0.1";
-    final int port = 10500;
+    final int port = testPortProvider.getFreePort();
 
     // create a dummy socket bound to a known port.
     try (ServerSocketChannel dummyServerSocket = ServerSocketChannel.open()) {
@@ -155,14 +145,14 @@ public class TestAvroSource {
     // As port is already in use, an exception is thrown and the source is stopped
     // cleaning up the opened sockets during source.start().
     Assert.assertEquals("Server is stopped", LifecycleState.STOP,
-            source.getLifecycleState());
+        source.getLifecycleState());
   }
 
   @Test
   public void testInvalidAddress()
       throws InterruptedException, IOException {
     final String invalidHost = "invalid.host";
-    final int port = 10501;
+    final int port = testPortProvider.getFreePort();
 
     Context context = new Context();
     context.put("port", String.valueOf(port));
@@ -232,28 +222,19 @@ public class TestAvroSource {
 
   private void doRequest(boolean serverEnableCompression, boolean clientEnableCompression,
                          int compressionLevel) throws InterruptedException, IOException {
-    boolean bound = false;
+    int selectedPort = testPortProvider.getFreePort();
 
-    for (int i = 0; i < 100 && !bound; i++) {
-      Context context = new Context();
-      context.put("port", String.valueOf(selectedPort = 41414 + i));
-      context.put("bind", "0.0.0.0");
-      context.put("threads", "50");
-      if (serverEnableCompression) {
-        context.put("compression-type", "deflate");
-      } else {
-        context.put("compression-type", "none");
-      }
-      Configurables.configure(source, context);
-      try {
-        source.start();
-        bound = true;
-      } catch (FlumeException e) {
-        /*
-         * NB: This assume the failure is to bind.
-         */
-      }
+    Context context = new Context();
+    context.put("port", String.valueOf(selectedPort));
+    context.put("bind", "0.0.0.0");
+    context.put("threads", "50");
+    if (serverEnableCompression) {
+      context.put("compression-type", "deflate");
+    } else {
+      context.put("compression-type", "none");
     }
+    Configurables.configure(source, context);
+    source.start();
 
     Assert
         .assertTrue("Reached start or error", LifecycleController.waitForOneOf(
@@ -311,7 +292,7 @@ public class TestAvroSource {
       NioClientSocketChannelFactory {
     private int compressionLevel;
 
-    public CompressionChannelFactory( int compressionLevel) {
+    public CompressionChannelFactory(int compressionLevel) {
       super();
       this.compressionLevel = compressionLevel;
     }
@@ -332,28 +313,19 @@ public class TestAvroSource {
 
   @Test
   public void testSslRequest() throws InterruptedException, IOException {
-    boolean bound = false;
 
-    for (int i = 0; i < 10 && !bound; i++) {
-      Context context = new Context();
+    Context context = new Context();
 
-      context.put("port", String.valueOf(selectedPort = 41414 + i));
-      context.put("bind", "0.0.0.0");
-      context.put("ssl", "true");
-      context.put("keystore", "src/test/resources/server.p12");
-      context.put("keystore-password", "password");
-      context.put("keystore-type", "PKCS12");
-      Configurables.configure(source, context);
-      try {
-        source.start();
-        bound = true;
-      } catch (FlumeException e) {
-        /*
-         * NB: This assume the failure is to bind.
-         */
-        Thread.sleep(100);
-      }
-    }
+
+    int selectedPort = testPortProvider.getFreePort();
+    context.put("port", String.valueOf(selectedPort));
+    context.put("bind", "0.0.0.0");
+    context.put("ssl", "true");
+    context.put("keystore", "src/test/resources/server.p12");
+    context.put("keystore-password", "password");
+    context.put("keystore-type", "PKCS12");
+    Configurables.configure(source, context);
+    source.start();
 
     Assert
         .assertTrue("Reached start or error", LifecycleController.waitForOneOf(
@@ -363,7 +335,7 @@ public class TestAvroSource {
 
     AvroSourceProtocol client = SpecificRequestor.getClient(
         AvroSourceProtocol.class, new NettyTransceiver(new InetSocketAddress(
-        selectedPort), new SSLChannelFactory()));
+            selectedPort), new SSLChannelFactory()));
 
     AvroFlumeEvent avroEvent = new AvroFlumeEvent();
 
@@ -407,7 +379,7 @@ public class TestAvroSource {
       try {
         SSLContext sslContext = SSLContext.getInstance("TLS");
         sslContext.init(null, new TrustManager[]{new PermissiveTrustManager()},
-                        null);
+            null);
         SSLEngine sslEngine = sslContext.createSSLEngine();
         sslEngine.setUseClientMode(true);
         // addFirst() will make SSL handling the first stage of decoding
@@ -479,7 +451,7 @@ public class TestAvroSource {
     doIpFilterTest(localhost, "deny:ip:127.0.0.2,allow:ip:*,deny:ip:" +
         localhost.getHostAddress(), true, false);
     doIpFilterTest(localhost, "deny:ip:" +
-        localhost.getHostAddress().substring(0, 3) + "*,allow:ip:*",
+            localhost.getHostAddress().substring(0, 3) + "*,allow:ip:*",
         false, false);
 
     // Private lambda expression to check the received FlumeException within this test
@@ -521,39 +493,28 @@ public class TestAvroSource {
   }
 
   public void doIpFilterTest(InetAddress dest, String ruleDefinition,
-      boolean eventShouldBeAllowed, boolean testWithSSL)
+                             boolean eventShouldBeAllowed, boolean testWithSSL)
       throws InterruptedException, IOException {
-    boolean bound = false;
 
-    for (int i = 0; i < 100 && !bound; i++) {
-
-      Context context = new Context();
-      context.put("port", String.valueOf(selectedPort = 41414 + i));
-      context.put("bind", "0.0.0.0");
-      context.put("ipFilter", "true");
-      if (ruleDefinition != null) {
-        context.put("ipFilterRules", ruleDefinition);
-      }
-      if (testWithSSL) {
-        logger.info("Client testWithSSL" + testWithSSL);
-        context.put("ssl", "true");
-        context.put("keystore", "src/test/resources/server.p12");
-        context.put("keystore-password", "password");
-        context.put("keystore-type", "PKCS12");
-      }
-      // Invalid configuration may result in a FlumeException
-      Configurables.configure(source, context);
-
-      try {
-        source.start();
-        bound = true;
-      } catch (FlumeException e) {
-        /*
-         * NB: This assume the failure is to bind.
-         */
-        Thread.sleep(100);
-      }
+    Context context = new Context();
+    int selectedPort = testPortProvider.getFreePort();
+    context.put("port", String.valueOf(selectedPort));
+    context.put("bind", "0.0.0.0");
+    context.put("ipFilter", "true");
+    if (ruleDefinition != null) {
+      context.put("ipFilterRules", ruleDefinition);
     }
+    if (testWithSSL) {
+      logger.info("Client testWithSSL" + testWithSSL);
+      context.put("ssl", "true");
+      context.put("keystore", "src/test/resources/server.p12");
+      context.put("keystore-password", "password");
+      context.put("keystore-type", "PKCS12");
+    }
+    // Invalid configuration may result in a FlumeException
+    Configurables.configure(source, context);
+
+    source.start();
 
     Assert
         .assertTrue("Reached start or error", LifecycleController.waitForOneOf(
