+++ flume-ng-embedded-agent/pom.xml
@@ -81,5 +81,11 @@ limitations under the License.
       <scope>test</scope>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 </project>
