+++ flume-ng-legacy-sources/flume-avro-source/pom.xml
@@ -135,6 +135,12 @@ limitations under the License.
       <artifactId>servlet-api</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
 </project>
