+++ flume-ng-legacy-sources/flume-thrift-source/pom.xml
@@ -170,6 +170,12 @@ limitations under the License.
       <artifactId>commons-lang</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
 </project>
