+++ flume-ng-sinks/flume-ng-kafka-sink/pom.xml
@@ -86,6 +86,12 @@
       <version>${kafka.version}</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.flume</groupId>
+      <artifactId>flume-test-common</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
 </project>
