+++ flume-ng-tests/pom.xml
@@ -71,7 +71,6 @@
       <plugin>    
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>${mvn-surefire-plugin.version}</version>
         <executions>
           <execution>
             <id>default-test</id>
